added pytest configuration and test for local Postgres connectivity #13 	deleted:    tests/integration/postgres_connectivity_test.py 	new file:   tests/integration/test_local_postgres.py
